Added a "strict JSON" mode to the text generator and compiler
This will add quotes around field names, as required by the official
standard. By default it will leave quotes out, as it is more readable,
more compact, and is accepted by almost all JSON parsers.
The -S switch to flatc turns on strict mode.

As per rfc 7159.

Tested: on Windows.

+++ src/idl_parser.cpp
@@ -669,7 +669,7 @@ void Parser::ParseDecl() {
   // Check if this is a table that has manual id assignments
   auto &fields = struct_def.fields.vec;
   if (!struct_def.fixed && fields.size()) {
-    int num_id_fields = 0;
+    size_t num_id_fields = 0;
     for (auto it = fields.begin(); it != fields.end(); ++it) {
       if ((*it)->attributes.Lookup("id")) num_id_fields++;
     }
